requirements.txt
configparser
mysql-connector
RPi.GPIO
spi